pyproject.toml
[tool.poetry]
name = "manager-mccode"
version = "0.1.0"
description = "Screen activity summarization using Gemini Vision API"
authors = ["Jake Gearon <jake.gearon@gmail.com>"]
readme = "README.md"
packages = [{include = "manager_mccode"}]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
pillow = "^10.2.0"  # for screenshot handling
google-generativeai = "^0.3.0"
pyautogui = "^0.9.54"  # for screenshots
pydantic = "^2.6.0"  # for structured output
duckdb = "^0.9.2"  # Add DuckDB
rich = "^13.7.0"
click = "^8.1.7"

[tool.poetry.scripts]
mccode = "manager_mccode.cli.service:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
